Query Datim SQLViews

Query Datim SQLViews

Usage

datim_sqlviews(
  username,
  password,
  view_name = NULL,
  dataset = FALSE,
  datauid = NULL,
  query = NULL,
  baseurl = NULL
)

Arguments

username

Datim username

password

Datim password

view_name

Datim SQLView name

dataset

Return SQLView dataset or uid? Default is false

datauid

Data UID

query

SQLView Query params, a list containing type and params key value pairs

baseurl

Datim API Base URL

Value

SQLView uid or dataset as data frame

Note

This function should be used to identify Datim SQLView and Extract Data

Examples

if (FALSE) { # \dontrun{
  library(grabr)

  datim_sqlviews(
    username =glamr::datim_user(),
    password =glamr::datim_pwd(),
    view_name = "A list of OUs",
    dataset = TRUE
  )
} # }